Reorganizing Files
Organized all the files into Activities, Adapters, and Data packages.

Deleted an older file Homescreen.kt which is not in use since HomeActivity.kt already exists.
