+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MaterialComponents.Light.NoActionBar">
-        <activity android:name=".HomeActivity" android:exported="true"/>
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".ManagerControlsActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".Activities.HomeActivity" android:exported="true"/>
+        <activity android:name=".Activities.RegisterActivity" />
+        <activity android:name=".Activities.ManagerControlsActivity"></activity>
+        <activity android:name=".Activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
