+++ app/src/main/java/com/proshiftteam/proshift/Activities/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.proshiftteam.proshift
+package com.proshiftteam.proshift.Activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.proshiftteam.proshift.R
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
