+++ app/src/main/java/com/proshiftteam/proshift/Activities/RegisterActivity.kt
@@ -1,8 +1,9 @@
-package com.proshiftteam.proshift
+package com.proshiftteam.proshift.Activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.proshiftteam.proshift.R
 
 class RegisterActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
