+++ app/src/main/java/com/proshiftteam/proshift/DataFiles/ScheduleData.kt
@@ -1,4 +1,4 @@
-package com.proshiftteam.proshift
+package com.proshiftteam.proshift.DataFiles
 
 /**
  * Hour Minute Second
