--- app/src/main/java/com/proshiftteam/proshift/HomeScreen.kt
@@ -1,7 +1,0 @@
-package com.proshiftteam.proshift
-
-import androidx.appcompat.app.AppCompatActivity
-
-class HomeScreen: AppCompatActivity() {
-
-}
