+++ app/src/main/res/layout/activity_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisterActivity">
+    tools:context=".Activities.RegisterActivity">
 
     <EditText
         android:id="@+id/plainTextRegisterFirstName"
